Moves ResourceList, ResourceStatusMarker
