+++ amundsen_application/static/js/components/ResourceList/PaginatedApiResourceList/index.spec.tsx
@@ -3,14 +3,12 @@
 
 import * as React from 'react';
 import Pagination from 'react-js-pagination';
-
 import { shallow } from 'enzyme';
 
 import { ResourceType } from 'interfaces';
-import ResourceListItem from 'components/common/ResourceListItem/index';
+import ResourceListItem from 'components/ResourceListItem/index';
 
 import PaginatedApiResourceList, { PaginatedApiResourceListProps } from '.';
-import * as CONSTANTS from '../constants';
 
 describe('PaginatedApiResourceList', () => {
   const setStateSpy = jest.spyOn(
