+++ amundsen_application/static/js/components/ResourceList/PaginatedResourceList/index.tsx
@@ -3,7 +3,8 @@
 
 import * as React from 'react';
 import Pagination from 'react-js-pagination';
-import ResourceListItem from 'components/common/ResourceListItem';
+
+import ResourceListItem from 'components/ResourceListItem';
 import { Resource } from 'interfaces';
 import * as Constants from '../constants';
