+++ amundsen_application/static/js/components/ResourceList/styles.scss
