+++ amundsen_application/static/js/components/ResourceListItem/SchemaInfo/index.tsx
