+++ amundsen_application/static/js/components/ResourceListItem/SchemaInfo/styles.scss
