+++ amundsen_application/static/js/components/ResourceListItem/TableListItem/index.spec.tsx
@@ -2,12 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as React from 'react';
-
 import { shallow } from 'enzyme';
 
 import { Link } from 'react-router-dom';
 import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
-import SchemaInfo from 'components/common/ResourceListItem/SchemaInfo';
+import SchemaInfo from 'components/ResourceListItem/SchemaInfo';
 import { ResourceType } from 'interfaces';
 
 import * as ConfigUtils from 'config/config-utils';
