+++ amundsen_application/static/js/components/ResourceListItem/styles.scss
