+++ amundsen_application/static/js/components/ResourceStatusMarker/styles.scss
