+++ amundsen_application/static/js/pages/SearchPage/index.spec.tsx
@@ -4,7 +4,6 @@
 import * as React from 'react';
 import * as DocumentTitle from 'react-document-title';
 import * as History from 'history';
-
 import { shallow } from 'enzyme';
 
 import { ResourceType } from 'interfaces';
@@ -12,7 +11,7 @@ import { ResourceType } from 'interfaces';
 import ResourceSelector from 'pages/SearchPage/ResourceSelector';
 import SearchFilter from 'pages/SearchPage/SearchFilter';
 import SearchPanel from 'pages/SearchPage/SearchPanel';
-import PaginatedApiResourceList from 'components/common/ResourceList/PaginatedApiResourceList';
+import PaginatedApiResourceList from 'components/ResourceList/PaginatedApiResourceList';
 
 import globalState from 'fixtures/globalState';
 import {
