+++ amundsen_application/static/js/pages/TableDetailPage/TableDashboardResourceList/index.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 
-import ResourceList from 'components/common/ResourceList';
+import ResourceList from 'components/ResourceList';
 import ShimmeringResourceLoader from 'components/common/ShimmeringResourceLoader';
 
 import { GlobalState } from 'ducks/rootReducer';
